% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/famnesia-package.R
\docType{package}
\name{famnesia-package}
\alias{famnesia}
\alias{famnesia-package}
\title{famnesia: Anonymising Familias Files}
\description{
A 'shiny' application for anonymising files exported from the 'Familias' software for forensic kinship analysis. Pedigrees, marker data, allele frequencies and mutation models can be masked or modified, with options for preserving likelihood ratios exactly. The application is built on the 'pedsuite' packages for pedigree analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/magnusdv/famnesia}
  \item Report bugs at \url{https://github.com/magnusdv/famnesia/issues}
}

}
\author{
\strong{Maintainer}: Magnus Dehli Vigeland \email{m.d.vigeland@medisin.uio.no} (\href{https://orcid.org/0000-0002-9134-4962}{ORCID})

Authors:
\itemize{
  \item Magnus Dehli Vigeland \email{m.d.vigeland@medisin.uio.no} (\href{https://orcid.org/0000-0002-9134-4962}{ORCID})
}

}
\keyword{internal}
