% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchApp.R
\name{launchApp}
\alias{launchApp}
\title{Launch Famnesia}
\usage{
launchApp(launch.browser = TRUE, ...)
}
\arguments{
\item{launch.browser}{Whether to launch the app in a browser. Default: TRUE.}

\item{...}{Further arguments passed to \code{\link[shiny:runApp]{shiny::runApp()}}, such as
\code{quiet}, \code{port} and \code{host}}
}
\value{
The value returned by \code{\link[shiny:runApp]{shiny::runApp()}}
}
\description{
This function launches the Famnesia Shiny application.
}
\examples{
if(interactive()) launchApp()

}
